added setup.py file and reorganized directory; pep-8-ified spacing in some files; fixed db stuff
