+++ bvp/defaults.cfg
@@ -1,6 +1,8 @@
 [db]
 dbname = bvp
 dbhost = http://host_ip:port
+is_verbose = True
+return_objects = True
 
 [path]
 blender_cmd = blender
