+++ bvp/options.cfg
@@ -1,6 +1,8 @@
 [db]
 dbname = bvp
-dbhost = http://peyote:5984
+dbhost = http://localhost:5984
+is_verbose = True
+return_objects = True
 
 [path]
 blender_cmd = blender
