+++ bvp/utils/__init__.py
@@ -1,0 +1,10 @@
+""" 
+Initialization of utils.
+Couple handy-dandy functions:
+"""
+from __future__ import absolute_import
+
+from . import basics
+from . import blender
+from . import bvpMath
+#from . import plot # ? necessary? NO. Kill me.
